+++ src/editors/templates/dotype-wizards.ts
@@ -22,23 +22,17 @@ import {
 import {
   addReferencedDataTypes,
   allDataTypeSelector,
-  buildListFromStringArray,
   CreateOptions,
-  functionalConstraintEnum,
-  predefinedBasicTypeEnum,
   unifyCreateActionArray,
   UpdateOptions,
-  valKindEnum,
   WizardOptions,
 } from './foundation.js';
 
 import { Select } from '@material/mwc-select';
-import {
-  SelectedEvent,
-  SingleSelectedEvent,
-} from '@material/mwc-list/mwc-list-foundation';
+import { SingleSelectedEvent } from '@material/mwc-list/mwc-list-foundation';
 import { ListItem } from '@material/mwc-list/mwc-list-item';
 import { List } from '@material/mwc-list';
+import { createDaWizard, editDAWizard } from '../../wizards/da.js';
 
 function updateSDoAction(element: Element): WizardActor {
   return (inputs: WizardInput[]): EditorAction[] => {
@@ -181,285 +175,6 @@ function sDOWizard(options: WizardOptions): Wizard | undefined {
                 >${dataType.id}</mwc-list-item
               >`
           )}</mwc-select
-        >`,
-      ],
-    },
-  ];
-}
-
-function updateDaAction(element: Element): WizardActor {
-  return (inputs: WizardInput[]): EditorAction[] => {
-    const name = getValue(inputs.find(i => i.label === 'name')!)!;
-    const desc = getValue(inputs.find(i => i.label === 'desc')!);
-    const bType = getValue(inputs.find(i => i.label === 'bType')!)!;
-    const type =
-      bType === 'Enum' || bType === 'Struct'
-        ? getValue(inputs.find(i => i.label === 'type')!)
-        : null;
-    const sAddr = getValue(inputs.find(i => i.label === 'sAddr')!);
-    const valKind =
-      getValue(inputs.find(i => i.label === 'valKind')!) !== ''
-        ? getValue(inputs.find(i => i.label === 'valKind')!)
-        : null;
-    const valImport =
-      getValue(inputs.find(i => i.label === 'valImport')!) !== ''
-        ? getValue(inputs.find(i => i.label === 'valImport')!)
-        : null;
-
-    const actions: EditorAction[] = [];
-    if (
-      name === element.getAttribute('name') &&
-      desc === element.getAttribute('desc') &&
-      bType === element.getAttribute('bType') &&
-      type === element.getAttribute('type') &&
-      sAddr === element.getAttribute('sAddr') &&
-      valKind === element.getAttribute('valKind') &&
-      valImport === element.getAttribute('valImprot')
-    ) {
-      return [];
-    }
-
-    const newElement = <Element>element.cloneNode(false);
-    newElement.setAttribute('name', name);
-    if (desc === null) newElement.removeAttribute('desc');
-    else newElement.setAttribute('desc', desc);
-    newElement.setAttribute('bType', bType);
-    if (type === null) newElement.removeAttribute('type');
-    else newElement.setAttribute('type', type);
-    if (sAddr === null) newElement.removeAttribute('sAddr');
-    else newElement.setAttribute('sAddr', sAddr);
-    if (valKind === null) newElement.removeAttribute('valKind');
-    else newElement.setAttribute('valKind', valKind);
-    if (valImport === null) newElement.removeAttribute('valImport');
-    else newElement.setAttribute('valImport', valImport);
-    actions.push({
-      old: { element },
-      new: { element: newElement },
-    });
-
-    return actions;
-  };
-}
-
-function createDaAction(parent: Element): WizardActor {
-  return (inputs: WizardInput[]): EditorAction[] => {
-    const name = getValue(inputs.find(i => i.label === 'name')!)!;
-    const desc = getValue(inputs.find(i => i.label === 'desc')!);
-    const bType = getValue(inputs.find(i => i.label === 'bType')!)!;
-    const type =
-      bType === 'Enum' || bType === 'Struct'
-        ? getValue(inputs.find(i => i.label === 'type')!)
-        : null;
-    const sAddr = getValue(inputs.find(i => i.label === 'sAddr')!);
-    const valKind =
-      getValue(inputs.find(i => i.label === 'valKind')!) !== ''
-        ? getValue(inputs.find(i => i.label === 'valKind')!)
-        : null;
-    const valImport =
-      getValue(inputs.find(i => i.label === 'valImport')!) !== ''
-        ? getValue(inputs.find(i => i.label === 'valImport')!)
-        : null;
-
-    const actions: EditorAction[] = [];
-
-    const element = createElement(parent.ownerDocument, 'DA', {
-      name,
-      desc,
-      bType,
-      type,
-      sAddr,
-      valKind,
-      valImport,
-    });
-    actions.push({
-      new: {
-        parent,
-        element,
-        reference: getReference(parent, <SCLTag>element.tagName),
-      },
-    });
-
-    return actions;
-  };
-}
-
-function dAWizard(options: WizardOptions): Wizard | undefined {
-  const doc = (<UpdateOptions>options).doc
-    ? (<UpdateOptions>options).doc
-    : (<CreateOptions>options).parent.ownerDocument;
-  const da =
-    Array.from(
-      doc.querySelectorAll(
-        selector('DA', (<UpdateOptions>options).identity ?? NaN)
-      )
-    ).find(isPublic) ?? null;
-
-  const [
-    title,
-    action,
-    type,
-    deleteButton,
-    name,
-    desc,
-    bTypeList,
-    sAddr,
-    valKindList,
-    valImportList,
-    fcList,
-  ] = da
-    ? [
-        get('da.wizard.title.edit'),
-        updateDaAction(da),
-        da.getAttribute('type'),
-        html`<mwc-button
-          icon="delete"
-          trailingIcon
-          label="${translate('delete')}"
-          @click=${(e: MouseEvent) => {
-            e.target!.dispatchEvent(newWizardEvent());
-            e.target!.dispatchEvent(
-              newActionEvent({
-                old: {
-                  parent: da.parentElement!,
-                  element: da,
-                  reference: da.nextSibling,
-                },
-              })
-            );
-          }}
-          fullwidth
-        ></mwc-button> `,
-        da.getAttribute('name'),
-        da.getAttribute('desc'),
-        buildListFromStringArray(
-          predefinedBasicTypeEnum,
-          da.getAttribute('bType')
-        ),
-        da.getAttribute('sAddr'),
-        buildListFromStringArray(valKindEnum, da.getAttribute('valKind')),
-        buildListFromStringArray(
-          [null, 'true', 'false'],
-          da.getAttribute('valImport')
-        ),
-        buildListFromStringArray(
-          functionalConstraintEnum,
-          da.getAttribute('fc')
-        ),
-      ]
-    : [
-        get('da.wizard.title.add'),
-        createDaAction((<CreateOptions>options).parent),
-        null,
-        html``,
-        '',
-        null,
-        buildListFromStringArray(predefinedBasicTypeEnum, 'Struct'),
-        null,
-        buildListFromStringArray(valKindEnum, null),
-        buildListFromStringArray([null, 'true', 'false'], null),
-        buildListFromStringArray(functionalConstraintEnum, null),
-      ];
-
-  const types = Array.from(doc.querySelectorAll('DAType, EnumType'))
-    .filter(isPublic)
-    .filter(type => type.getAttribute('id'));
-
-  return [
-    {
-      title,
-      element: da ?? undefined,
-      primary: { icon: '', label: get('save'), action: action },
-      content: [
-        deleteButton,
-        html`<wizard-textfield
-          label="name"
-          .maybeValue=${name}
-          helper="${translate('scl.name')}"
-          required
-          pattern="${patterns.alphanumericFirstLowerCase}"
-          dialogInitialFocus
-        >
-          ></wizard-textfield
-        >`,
-        html`<wizard-textfield
-          label="desc"
-          helper="${translate('scl.desc')}"
-          .maybeValue=${desc}
-          nullable
-          pattern="${patterns.normalizedString}"
-        ></wizard-textfield>`,
-        html`<mwc-select
-          fixedMenuPosition
-          label="bType"
-          helper="${translate('scl.bType')}"
-          required
-          @selected=${(e: SelectedEvent) => {
-            const bTypeOriginal = da?.getAttribute('bType') ?? '';
-            const bType = (<Select>e.target).selected!.value!;
-
-            const typeUI = <Select>(
-              (<Select>e.target).parentElement!.querySelector(
-                'mwc-select[label="type"]'
-              )!
-            );
-
-            Array.from(typeUI.children).forEach(child => {
-              (<ListItem>child).disabled = !child.classList.contains(bType);
-              (<ListItem>child).noninteractive =
-                !child.classList.contains(bType);
-              (<ListItem>child).style.display = !child.classList.contains(bType)
-                ? 'none'
-                : '';
-              (<ListItem>child).selected =
-                bTypeOriginal === bType
-                  ? (<ListItem>child).value === type
-                  : child.classList.contains(bType);
-            });
-
-            typeUI.disabled = !(bType === 'Enum' || bType === 'Struct');
-            typeUI.requestUpdate();
-          }}
-          >${bTypeList}</mwc-select
-        >`,
-        html`<mwc-select
-          fixedMenuPosition
-          label="type"
-          helper="${translate('scl.type')}"
-          >${types.map(
-            dataType =>
-              html`<mwc-list-item
-                class="${dataType.tagName === 'EnumType' ? 'Enum' : 'Struct'}"
-                value=${dataType.id}
-                ?selected=${dataType.id === type}
-                >${dataType.id}</mwc-list-item
-              >`
-          )}</mwc-select
-        >`,
-        html`<wizard-textfield
-          label="sAddr"
-          helper="${translate('scl.sAddr')}"
-          .maybeValue=${sAddr}
-          nullable
-          pattern="${patterns.normalizedString}"
-        ></wizard-textfield>`,
-        html`<mwc-select
-          label="valKind"
-          helper="${translate('scl.valKind')}"
-          fixedMenuPosition
-          >${valKindList}</mwc-select
-        >`,
-        html`<mwc-select
-          fixedMenuPosition
-          label="valImport"
-          helper="${translate('scl.valImport')}"
-          >${valImportList}</mwc-select
-        >`,
-        html`<mwc-select
-          label="fc"
-          helper="${translate('scl.fc')}"
-          required
-          fixedMenuPosition
-          >${fcList}</mwc-select
         >`,
       ],
     },
@@ -670,10 +385,8 @@ export function dOTypeWizard(
             trailingIcon
             label="${translate('scl.DA')}"
             @click=${(e: Event) => {
-              const wizard = dAWizard({
-                parent: dotype,
-              });
-              if (wizard) e.target!.dispatchEvent(newWizardEvent(wizard));
+              if (dotype)
+                e.target!.dispatchEvent(newWizardEvent(createDaWizard(dotype)));
               e.target!.dispatchEvent(newWizardEvent());
             }}
           ></mwc-button>
@@ -684,11 +397,13 @@ export function dOTypeWizard(
             @selected=${(e: SingleSelectedEvent) => {
               const item = <ListItem>(<List>e.target).selected;
 
+              const daIdentity = (<ListItem>(<List>e.target).selected).value;
+              const da = doc.querySelector(selector('DA', daIdentity));
+
               const wizard = item.classList.contains('DA')
-                ? dAWizard({
-                    identity: item.value,
-                    doc,
-                  })
+                ? da
+                  ? editDAWizard(da)
+                  : undefined
                 : sDOWizard({
                     identity: item.value,
                     doc,
