+++ src/editors/templates/foundation.ts
@@ -121,61 +121,6 @@ export function buildListFromStringArray(
   );
 }
 
-export const predefinedBasicTypeEnum = [
-  'BOOLEAN',
-  'INT8',
-  'INT16',
-  'INT24',
-  'INT32',
-  'INT64',
-  'INT128',
-  'INT8U',
-  'INT16U',
-  'INT24U',
-  'INT32U',
-  'FLOAT32',
-  'FLOAT64',
-  'Enum',
-  'Dbpos',
-  'Tcmd',
-  'Quality',
-  'Timestamp',
-  'VisString32',
-  'VisString64',
-  'VisString65',
-  'VisString129',
-  'VisString255',
-  'Octet64',
-  'Unicode255',
-  'Struct',
-  'EntryTime',
-  'Check',
-  'ObjRef',
-  'Currency',
-  'PhyComAddr',
-  'TrgOps',
-  'OptFlds',
-  'SvOptFlds',
-  'EntryID',
-];
-
-export const functionalConstraintEnum = [
-  'ST',
-  'MX',
-  'SP',
-  'SV',
-  'CF',
-  'DC',
-  'SG',
-  'SE',
-  'SR',
-  'OR',
-  'BL',
-  'EX',
-];
-
-export const valKindEnum = [null, 'Spec', 'Conf', 'RO', 'Set'];
-
 /** Common `CSS` styles used by DataTypeTemplate subeditors */
 export const styles = css`
   :host(.moving) section {
