+++ src/open-scd.ts
@@ -34,6 +34,7 @@ import '@material/mwc-top-app-bar-fixed';
 import './filtered-list.js';
 import './wizard-dialog.js';
 import './wizard-textfield.js';
+import './wizard-select.js';
 import './finder-pane.js';
 
 import { newOpenDocEvent, newPendingStateEvent } from './foundation.js';
