+++ test/integration/editors/templates/dotype-wizarding.test.ts
@@ -168,224 +168,6 @@ describe('DOType wizards', () => {
     });
   });
 
-  describe('defines a dAWizard to edit an existing DA', () => {
-    let nameField: WizardTextField;
-    let primayAction: HTMLElement;
-    let deleteButton: HTMLElement;
-    let bTypeSelect: Select;
-    let typeSelect: Select;
-
-    beforeEach(async () => {
-      (<ListItem>(
-        dOTypeList.querySelector('mwc-list-item[value="#Dummy.LLN0.Mod"]')
-      )).click();
-      await parent.requestUpdate();
-      await new Promise(resolve => setTimeout(resolve, 100)); // await animation
-      (<HTMLElement>(
-        parent.wizardUI?.dialog?.querySelector(
-          'mwc-list-item[value="#Dummy.LLN0.Mod>stVal"]'
-        )
-      )).click();
-      await parent.requestUpdate();
-      await new Promise(resolve => setTimeout(resolve, 100)); // await animation
-
-      nameField = <WizardTextField>(
-        parent.wizardUI.dialog?.querySelector('wizard-textfield[label="name"]')
-      );
-      primayAction = <HTMLElement>(
-        parent.wizardUI.dialog?.querySelector(
-          'mwc-button[slot="primaryAction"]'
-        )
-      );
-      deleteButton = <HTMLElement>(
-        parent.wizardUI.dialog?.querySelector('mwc-button[icon="delete"]')
-      );
-      bTypeSelect = <Select>(
-        parent.wizardUI.dialog?.querySelector('mwc-select[label="bType"]')
-      );
-      typeSelect = <Select>(
-        parent.wizardUI.dialog?.querySelector('mwc-select[label="type"]')
-      );
-    });
-
-    it('looks like the latest snapshot', () => {
-      expect(parent.wizardUI.dialog).to.equalSnapshot();
-    });
-    it('edits DA attributes name', async () => {
-      expect(
-        doc.querySelector('DOType[id="Dummy.LLN0.Mod"] > DA[name="stVal"]')
-      ).to.exist;
-      nameField.value = 'newCtlVal';
-      await parent.requestUpdate();
-      primayAction.click();
-      await parent.requestUpdate();
-      expect(
-        doc.querySelector('DOType[id="Dummy.LLN0.Mod"] > DA[name="stVal"]')
-      ).to.not.exist;
-      expect(
-        doc.querySelector('DOType[id="Dummy.LLN0.Mod"] > DA[name="newCtlVal"]')
-      ).to.exist;
-    });
-    it('deletes the DA element on delete button click', async () => {
-      expect(
-        doc.querySelector('DOType[id="Dummy.LLN0.Mod"] > DA[name="stVal"]')
-      ).to.exist;
-      expect(
-        doc.querySelectorAll('DOType[id="Dummy.LLN0.Mod"] > DA').length
-      ).to.equal(14);
-      deleteButton.click();
-      await parent.requestUpdate();
-      expect(
-        doc.querySelector('DOType[id="Dummy.LLN0.Mod"] > DA[name="stVal"]')
-      ).to.not.exist;
-      expect(
-        doc.querySelectorAll('DOType[id="Dummy.LLN0.Mod"] > DA').length
-      ).to.equal(13);
-    });
-    it('does not edit DA element without changes', async () => {
-      const originData = (<Element>(
-        doc.querySelector('DOType[id="Dummy.LLN0.Mod"] > DA[name="stVal"]')
-      )).cloneNode(true);
-      primayAction.click();
-      await parent.requestUpdate();
-      expect(
-        originData.isEqualNode(
-          doc.querySelector('DOType[id="Dummy.LLN0.Mod"] > DA[name="stVal"]')
-        )
-      ).to.be.true;
-    });
-    it('filters the type selector to EnumTypes if bType is Enum', () => {
-      expect(
-        Array.from(typeSelect!.querySelectorAll('mwc-list-item')).find(
-          item =>
-            (<ListItem>item).className === 'Enum' &&
-            (<ListItem>item).style.display === 'none'
-        )
-      ).to.be.undefined;
-    });
-    it('filters the type selector to DATypes if bType is Struct', async () => {
-      bTypeSelect.value = 'Struct';
-      await parent.requestUpdate();
-      expect(
-        Array.from(typeSelect!.querySelectorAll('mwc-list-item')).find(
-          item =>
-            (<ListItem>item).className === 'Struct' &&
-            (<ListItem>item).style.display === 'none'
-        )
-      ).to.be.undefined;
-    });
-    it('disables the type selector if bType is not Enum nor Struct', async () => {
-      expect(typeSelect.disabled).to.be.false;
-      bTypeSelect.value = 'BOOLEAN';
-      await parent.requestUpdate();
-      expect(typeSelect.disabled).to.be.true;
-    });
-  });
-
-  describe('defines a dAWizard to create a new DA element', () => {
-    let nameField: WizardTextField;
-    let descField: WizardTextField;
-    let sAddrField: WizardTextField;
-    let bTypeSelect: Select;
-    let valKindSelect: Select;
-    let valImportSelect: Select;
-    let fcSelect: Select;
-    let primayAction: HTMLElement;
-
-    beforeEach(async () => {
-      (<ListItem>(
-        dOTypeList.querySelector('mwc-list-item[value="#Dummy.LLN0.Mod"]')
-      )).click();
-      await parent.requestUpdate();
-      await new Promise(resolve => setTimeout(resolve, 100)); // await animation
-      (<HTMLElement>(
-        parent.wizardUI.dialog?.querySelectorAll(
-          'mwc-button[icon="playlist_add"]'
-        )[1]
-      )).click();
-      await parent.requestUpdate();
-      await new Promise(resolve => setTimeout(resolve, 100)); // await animation
-
-      nameField = <WizardTextField>(
-        parent.wizardUI.dialog?.querySelector('wizard-textfield[label="name"]')
-      );
-      descField = <WizardTextField>(
-        parent.wizardUI.dialog?.querySelector('wizard-textfield[label="desc"]')
-      );
-      sAddrField = <WizardTextField>(
-        parent.wizardUI.dialog?.querySelector('wizard-textfield[label="sAddr"]')
-      );
-      bTypeSelect = <Select>(
-        parent.wizardUI.dialog?.querySelector('mwc-select[label="bType"]')
-      );
-      valKindSelect = <Select>(
-        parent.wizardUI.dialog?.querySelector('mwc-select[label="valKind"]')
-      );
-      valImportSelect = <Select>(
-        parent.wizardUI.dialog?.querySelector('mwc-select[label="valImport"]')
-      );
-      fcSelect = <Select>(
-        parent.wizardUI.dialog?.querySelector('mwc-select[label="fc"]')
-      );
-      primayAction = <HTMLElement>(
-        parent.wizardUI.dialog?.querySelector(
-          'mwc-button[slot="primaryAction"]'
-        )
-      );
-    });
-
-    it('looks like the latest snapshot', () => {
-      expect(parent.wizardUI.dialog).to.equalSnapshot();
-    });
-    it('creates a new DA element', async () => {
-      expect(
-        doc.querySelector(
-          'DOType[id="Dummy.LLN0.Mod"] > DA[name="newDAElement"]'
-        )
-      ).to.not.exist;
-      nameField.value = 'newDAElement';
-      fcSelect.value = 'ST';
-      await parent.requestUpdate();
-      primayAction.click();
-      await parent.requestUpdate();
-      expect(
-        doc.querySelector(
-          'DOType[id="Dummy.LLN0.Mod"] > DA[name="newDAElement"]:not([desc]):not([sAddr])[bType="Struct"][type="Dummy.Units"]:not([valKind]):not([valImport])'
-        )
-      ).to.exist;
-    });
-    it('creates yet another new DA element', async () => {
-      const name = 'newDAElement2';
-      const desc = 'newDAdesc';
-      const sAddr = 'myNewAddr';
-
-      expect(
-        doc.querySelector(
-          'DAType[id="#Dummy.LLN0.Mod"] > DA[name="newDAElement2"]'
-        )
-      ).to.not.exist;
-      nameField.value = name;
-      descField.nullable = false;
-      descField.value = desc;
-      sAddrField.nullable = false;
-      sAddrField.value = sAddr;
-      bTypeSelect.value = 'BOOLEAN';
-      valKindSelect.value = 'RO';
-      valImportSelect.value = 'true';
-      fcSelect.value = 'ST';
-
-      await parent.requestUpdate();
-      primayAction.click();
-      await parent.requestUpdate();
-      expect(
-        doc.querySelector(
-          `DOType[id="Dummy.LLN0.Mod"] >` +
-            `DA[name="${name}"][desc="${desc}"][sAddr="${sAddr}"][bType="BOOLEAN"]:not([type])[valKind="RO"][valImport="true"]`
-        )
-      ).to.exist;
-    });
-  });
-
   describe('defines a sDOWizard to edit an existing SDO', () => {
     let nameField: WizardTextField;
     let primayAction: HTMLElement;
